toke.c: One less token for missing format args
In commit 705fe0e5f8a, when I made the parser understand format syntax
itself, I had to add special handling for a terminating dot where for-
mat arguments were expected:

format =
@
.

The parser expects every format argument line to look like this:

formarg	:	/* NULL */
			{ $$ = NULL; }
	|	FORMLBRACK stmtseq FORMRBRACK
			{ $$ = op_unscope($2); }
	;

When the line break is encountered after the @, the FORMLBRACK token
is emitted, and the lexer switches into ‘normal’ (as opposed to for-
mat picture) mode.  When the final dot is encountered, since the
FORMLBRACK has already been emitted, the lexer has to conjure up a
FORMRBRACK as well, to avoid a syntax error.

I had it producing a semicolon before the FORMRBRACK, but that is not
necessary, because stmtseq can be null.  So this commit removes it.