regcomp.c: Make sure counter same in passes 1 and 2
The number of elements was not being incremented in pass 1, whereas that
number is needed later on in pass 1.  This did not cause a
bug, as currently, in pass 1 we care only if the count is 1 or not, and
this occurred only in a case where it would get incremented properly to
more than 1 anyway.  But this is a potential bug that should be
squelched before it happens.