More tweaks to Maintainers.pl
- One CGI test is currently customized
- Nine Text::Balanced tests are currently customized
- One Test::Harness test is currently customized and another is renamed,
requiring a MAP; plus all the module version numbers are bumped
- Getopt::Long doesn't require a MAP

I have requested new CPAN versions of various CPAN distros that are
currently customized with changes which should be merged back upstream
with a view to getting them back in sync as they are intended to be.
Namely: CGI, Sys-Syslog, Test-Harness, Text-Balanced and Text-Tabs+Wrap.