[perl #114040] Allow pod in quoted constructs
When the case = in toke.c:yylex is reached and PL_lex_state is
not LEX_NORMAL, that means we are in some sort of quoted construct,
and the entire construct’s content is in the current line buffer
(which, consequently contains more than one line).  So we need to check
that when encountering pod.  Quoted constructs need to be treated the
same way as string eval, which also puts all the code in the line
buffer.