Consistent unixy path handling in File::Find::_find_opt.
Back in a1ccf0c4149b we converted the current working directory to
Unix format on VMS, but neglected to change what later gets pasted
onto it with a hard-coded slash delimiter.  The resulting mongrel
filespec was invalid and of course would not appear to exist even
if the file did exist under a properly assembled name.

So this commit makes the use of Unix-style paths on VMS within
_find_opt consistent.

The bug was tickled by a recent change to Module::Pluggable, whose
tests and the tests of other modules that depend on it started
failing en masse.